Move Collection to metadata package
